refactor: ChartRenderer→ChartContentにリネームしTableContentを分離
ResultView.tsxからChartContent/TableContentを各ファイルに切り出し、
ResultViewをstate管理と切替ロジックに集中させた。
