+++ src/components/aggregation/Toolbar.tsx
@@ -1,7 +1,7 @@
 import type { AggResult } from "../../lib/agg/aggregate";
 import { downloadAllCSV } from "../../lib/agg/download";
 import { t } from "../../lib/i18n";
-import type { GtChartType } from "./ChartRenderer";
+import type { GtChartType } from "./ChartContent";
 import { ToggleButton, ToggleGroup } from "../shared/ToggleButton";
 import { useAggregation } from "./AggregationContext";
